Added followşng controller

+++ PsikoterapsitlerBurada/Controllers/API/FollwingController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNet.Identity;
+﻿using System.Linq;
+using Microsoft.AspNet.Identity;
 using PsikoterapsitlerBurada.Models;
 using System.Web.Http;
 
@@ -26,6 +27,7 @@ namespace PsikoterapsitlerBurada.Controllers.API
             };
 
             _context.Followings.Add(following);
+
             _context.SaveChanges();
 
             return Ok();
@@ -35,17 +37,14 @@ namespace PsikoterapsitlerBurada.Controllers.API
         public IHttpActionResult UnFollow(string id)
         {
             var userId = User.Identity.GetUserId();
-
-            var following = new Following()
-            {
-                FolloweeId = id,
-                FollowerId = userId
-            };
+            var following = _context.Followings.SingleOrDefault(f => f.FollowerId == userId && f.FolloweeId == id);
+            if (following == null) return BadRequest();
 
             _context.Followings.Remove(following);
             _context.SaveChanges();
 
             return Ok();
+
         }
     }
 }
